+++ Lumos/Source/Lumos/Scene/Scene.cpp
@@ -131,7 +131,7 @@ namespace Lumos
         for(auto entity : animatedSpriteView)
         {
             auto& animSprite = entity.GetComponent<Graphics::AnimatedSprite>();
-            animSprite.OnUpdate(timeStep.GetSeconds());
+            animSprite.OnUpdate((float)timeStep.GetSeconds());
         }
     }
